revert changes

+++ pkg/etcdutil/etcdutil_test.go
@@ -123,7 +123,7 @@ func (s *testEtcdutilSuite) TestMemberHelpers(c *C) {
 	// Test CheckClusterID
 	urlmap, err := types.NewURLsMap(cfg2.InitialCluster)
 	c.Assert(err, IsNil)
-	err = CheckClusterID(uint64(etcd1.Server.Cluster().ID()), urlmap, &tls.Config{})
+	err = CheckClusterID(etcd1.Server.Cluster().ID(), urlmap, &tls.Config{})
 	c.Assert(err, IsNil)
 
 	// Test RemoveEtcdMember
